text-extractor/requirements.txt
streamlit
pytesseract
opencv-python
numpy
Pillow
python-docx
fpdf
requests
easyocr